Refactor the input architecture.
This patch makes it simple to add more inputs and more decoders.
It also removes the generics from the pipeline in favor of simple enums.
The implementation of the RTP receiver was also moved to the
`compositor_pipeline` crate.

A big fault of this patch is that it doesn't address attaching multiple
decoders to a single input, but this requires further discussion.